Add ParquetBufferWriter for synchronous in-memory writing
Adds ParquetBufferWriter, a synchronous writer that outputs to a buffer.
Moves encoding functions to a separate src/encoding.ts file to improve
browser compatibility by avoiding Node.js-specific dependencies in core
encoding logic. Exports row group and footer encoding utilities and
includes comprehensive tests for the new writer.

#9

Plan: parquet-buffer-writer
